Make old-school ImageJ2 plugins actually work
My testbed here was "Add Noise...". Like many such older plugins, it
needs an ImageDisplay to operate upon, and the ImageDisplay is a BOTH
parameter, meaning it mutates in place.

So while this commit makes it work, napari-imagej doesn't understand
BOTH parameters yet, so a second copy of the same image appears as a
new layer, and the original layer also changes, unfortunately.