Add python <-> java converters in __init.py__
The reasoning is that we want these converters automatically available from napari-imagej whenever we want to convert. The options are then to:

1. Write this line globally, as it is, and then have to write in every file:

import napari_imagej._napari_converters

This is dumb, as we'd have to do it even if we called no functions from it.

2. Write this line in ImageJWidget. This is dumb, because that would require we do the import whenever we don't use the widget. This is mainly tests, but leaves room for action-at-a-distance errors.

3. Write this in __init__.py. It sucks to import scyjava.when_jvm_starts and napari_imagej._napari_converters.init_napari_converters automatically.

I think that 3 is the best of the three options, so it is what I did.