Switch layer transfer to the active layer
Well, sometimes. What I've actually done is provide a setting that lets
the user choose whether to just use the active layer or to prompt the
user to choose the layer to be transferred.

I figured this was the best middle ground between the two, especially
since I had implementations of both coded.