Restrict settings order in popup
Due to the way confuse maintains settings changes, we can get different
settings orderings in the popup as a function of the previous edits
made.

By editing over the settings present in the DEFAULT source
(src/napari-imagej/config-default.yaml), we get two nice effects:
1. Settings will always appear in the order listed in the default
configuration file
2. Defined settings that are NOT in the default configuration file are
not listed in this popup. This could happen if the user hacked their own
configuration, for example.

Unfortunately, this does not fix the way that settings changes are
dumped into the user's settings. Thus the settings can still appear out
of order in the user's file. More work needed on that one...