Remove Ops version pin
Unfortunately, this Ops pin seems to be the cause behind a crash of
imagej2 and napari when the ImageJ GUI button is clicked.

By removing the version pin, we will not have any of the functionality
that was incorporated in ops 0.49.0, requiring we comment out a couple
of tests. Once we have a new ImageJ2, we can uncomment those tests