Use existing imagej.gateway if already running
In a future PyImageJ release, when running in GUI mode (and
*only* when running in GUI mode), the created ImageJ2 gateway will
be cached in the global variable imagej.gateway, so that it is
accessible from other threads, since the imagej.init function
never returns in that case. We can utilize this in napari-imagej
to hook into that ImageJ2 gateway, rather than creating a new one.

By doing this, napari-imagej becomes usable interactively while
PyImageJ is in GUI mode, which means it can now work with a GUI on
macOS! But you need to start napari + PyImageJ in a special way:

  import imagej, napari
  napari.Viewer()
  imagej.init(mode='gui')

As an aside: this construction does *not* work on non-macOS
platforms, because imagej.init(mode='gui') manually blocks the
thread with a polling sleep loop, which means that napari.run()
does not get invoked to start up napari's event loop, which
results in napari hanging. However, on Linux (and likely Windows)
you can instead do:

  import imagej, napari
  imagej.gateway = imagej.init(mode='interactive')
  imagej.gateway.ui().showUI()
  napari.Viewer()
  napari.run()

To utilize the pathway enabled by this commit.