Start napari before ImageJ on MacOS tests
We run into https://github.com/imagej/pyimagej/issues/197
on MacOS if we don't start napari before we start ImageJ.
Since we want ImageJ to run with a session scope (because it
takes forever to start ImageJ, so only do it once), this means
we need a napari Viewer before any tests run.

Many tests also need a Viewer. But you're supposed to make that
Viewer with napari's make_napari_viewer fixture. THAT fixture
has object scope, so it makes a new Viewer for every test.

And that fixture complains if there is a Viewer that already
exists. :(

SO on MacOS, we make a viewer so we can start ImageJ. And then
dispose of it immediately.