Switch jvm_mode setting to enable_imagej_gui bool
This is simpler and easier for users to understand. We also label it
"Enable ImageJ GUI if possible" in the settings dialog, to give the hint
that in some scenarios it might not be possible to have a GUI. The logic
to explain about macOS limitations when the ImageJ GUI button is clicked
remains in place, so users still receive an explanation of this issue.
But when changing the toggle in the settings, it now just works, rather
than proactively preventing them from enabling it on macOS. This is
important because in the future, we hope to enable GUI mode on macOS
when launching napari-imagej via a dedicated napari-imagej CLI entry
point, even if it cannot work in a more natural way as other platforms.