Simplify validation logic
There is really no situation where we want to let the validation
logic change the value from an invalid to a valid value. Nor is there
a situation where we want to fail to assign new settings. Rather, we
just want to provide a pathway where validation problems can be
detected. This change unifies the validation logic into a single
validate() function that simply raises ValueError with an arg string
description in its args list for each validation issue noticed.

It is up to downstream calling code to invoke this function (or not)
if it cares about whether the settings might have problems.