Name typing_extensions module correctly
It looks like conda-forge has both typing_extensions (underscore) *and*
typing-extensions (dash), and they are the same? And both packages are
brought in as dependencies, regardless of which one we declare here.
Something inconsistent across upstream dependencies, perhaps?

But the actual feedstock repository is named typing_extensions-feedstock
(https://github.com/conda-forge/typing_extensions-feedstock), whereas
the naming with a dash does not exist, which is a strong hint that the
underscore version is preferred.